composer: remove the recording lock entirely + move Discard into the action row (#1245)
Deletes the swipe-up-to-lock hands-free gesture (#585), recordingLocked state,
lockRecording, autoLockPending and all lock UI (D22 hard-cut) — recording is now
hands-free by default. Discard moves into a single [Discard · Insert · Send]
action row. Mic tap now fires via a real .clickable on MicTriggerButton (the
deleted gesture used to service the tap). On-device journey proves tap-to-record.

Closes #1245
