+++ app/src/androidTest/java/com/pocketshell/app/composer/PromptComposerRecordingRowFitProofTest.kt
@@ -36,58 +36,44 @@ import org.junit.Test
 import org.junit.runner.RunWith
 
 /**
- * Issue #1152 — the recording-not-locked composer bottom row overflowed its
- * width and CLIPPED `Send` off the right edge (the cyan "S/e" sliver the
- * maintainer circled — audit D1). The old single row packed the editing-tools
- * group + Lock + Insert + Send onto one line; the content (~421dp of pills + gaps)
- * exceeded a phone's usable width, and with no scroll/weight the trailing `Send`
- * fell off the right edge.
+ * Issue #1152 — the recording composer bottom row overflowed its width and
+ * CLIPPED `Send` off the right edge (the cyan "S/e" sliver the maintainer circled
+ * — audit D1). Issue #1245 removed the hands-free Lock ENTIRELY (both the pill
+ * and, later, the inline toggle) and moved Discard onto the SAME balanced action
+ * row as Insert and Send. The recording bottom row is now a single right-aligned
+ * cluster of exactly three pills — [Discard · Insert · Send] — with no editing
+ * tools competing for that row (attach/snippets/slash are not shown mid-dictation,
+ * so Discard sits right next to Insert and Send).
  *
- * The maintainer's directive is FIT EVERYTHING — never hide a control to make
- * room. The fix relays the recording states into a TWO-ROW layout (secondary row:
- * tools + Discard; primary row: Insert + Send) so EVERY control fits with the
- * editing tools still mounted. (Issue #1245 later moved the hands-free Lock UP to
- * the recording/waveform row as an inline toggle, so it is no longer a
- * bottom-cluster pill; this test guards that the inline toggle also stays within
- * the band.)
+ * ## What this proves (regression guard, class-covering)
  *
- * ## What this proves (red → green, class-covering)
- *
- * It composes the PRODUCTION [SheetContent] in the recording state — both
- * NOT-LOCKED and LOCKED — pinned to a FIXED phone-width band, keyboard-UP (the
- * #780 synthetic-inset model), at font scale 1.0 AND 1.3 (the widest-content
- * case, since this is a width-overflow bug). It then asserts, for every
- * recording-row pill (Discard / Insert / Send) AND the editing-tools
- * group:
+ * It composes the PRODUCTION [SheetContent] in the recording state, pinned to a
+ * FIXED phone-width band, keyboard-UP (the #780 synthetic-inset model), at font
+ * scale 1.0 AND 1.3 (the widest-content case, since #1152 was a width-overflow
+ * bug). It then asserts, for every recording-row pill (Discard / Insert / Send):
  *  1. the pill is fully within the band horizontally (no edge spill), AND
  *  2. the pill is NOT CRUSHED — its width stays at/above a not-a-sliver floor.
  *
- * Both are needed because at a real phone width the single-row overflow does not
- * push `Send` cleanly off the edge — the un-weighted Row squeezes the trailing
- * `Send` down to a ~33dp sliver AT the band edge (the maintainer's cyan "S/e";
- * base evidence: `Send widthDp=33`). A bare containment / `assertIsDisplayed()`
- * still passes on that crushed sliver — the exact #657/F1 trap — so the
- * not-crushed floor is the LOAD-BEARING assertion that goes RED on base and GREEN
- * with the two-row fit (`Send widthDp≈90`).
+ * Both are needed because a single un-weighted overflow squeezes the trailing
+ * `Send` down to a sliver AT the band edge; a bare containment / `assertIsDisplayed()`
+ * still passes on that crushed sliver (the #657/F1 trap), so the not-crushed floor
+ * is the LOAD-BEARING assertion.
  *
  * ## Why the band, not just [assertNodeFullyWithinRoot]
  *
- * The overflow at a phone width is ~20–30dp — razor-thin against the DEVICE
- * screen, so a slightly-wider CI AVD would mask it if we measured only against
- * `onRoot()` (the whole screen). We therefore host [SheetContent] in a
- * `requiredWidth` band anchored top-START and measure containment RELATIVE to the
- * band, so the red is DETERMINISTIC on every AVD regardless of physical width
- * (the #813 model). [assertNodeFullyWithinRoot] is asserted too as an additive
- * "never off the whole screen" guard, but the band-relative check is the
- * load-bearing red→green.
+ * The overflow at a phone width is razor-thin against the DEVICE screen, so a
+ * slightly-wider CI AVD would mask it if we measured only against `onRoot()`. We
+ * therefore host [SheetContent] in a `requiredWidth` band anchored top-START and
+ * measure containment RELATIVE to the band, so the check is DETERMINISTIC on every
+ * AVD regardless of physical width (the #813 model). [assertNodeFullyWithinRoot]
+ * is asserted too as an additive "never off the whole screen" guard.
  *
  * ## Determinism
  *
- * Band width + height + the synthetic IME/nav/status insets are all fixed
- * synthetic inputs, so the geometry is identical on the dev-box AVD and CI
- * swiftshader. The synthetic ime() inset is HARD-asserted to have applied before
- * any geometry is judged (no `assumeTrue` skip — the #780/#736 rule): the row is
- * proven horizontally contained with the keyboard UP.
+ * Band width + height + the synthetic IME/nav/status insets are all fixed synthetic
+ * inputs, so the geometry is identical on the dev-box AVD and CI swiftshader. The
+ * synthetic ime() inset is HARD-asserted to have applied before any geometry is
+ * judged (no `assumeTrue` skip — the #780/#736 rule).
  */
 @RunWith(AndroidJUnit4::class)
 class PromptComposerRecordingRowFitProofTest {
@@ -99,28 +85,18 @@ class PromptComposerRecordingRowFitProofTest {
     private val observedNavBottomPx = mutableStateOf(0)
 
     @Test
-    fun recordingNotLockedRowPillsFullyWithinBand_phoneWidth() {
-        assertRecordingRowFits(locked = false, fontScale = 1.0f)
+    fun recordingRowPillsFullyWithinBand_phoneWidth() {
+        assertRecordingRowFits(fontScale = 1.0f)
     }
 
     @Test
-    fun recordingNotLockedRowPillsFullyWithinBand_largeFont() {
+    fun recordingRowPillsFullyWithinBand_largeFont() {
         // 1.3 is the widest-content case — this is a width-overflow bug, so cover
         // the class (largest font) not just the reported 1.0 instance.
-        assertRecordingRowFits(locked = false, fontScale = 1.3f)
+        assertRecordingRowFits(fontScale = 1.3f)
     }
 
-    @Test
-    fun recordingLockedRowPillsFullyWithinBand_phoneWidth() {
-        assertRecordingRowFits(locked = true, fontScale = 1.0f)
-    }
-
-    @Test
-    fun recordingLockedRowPillsFullyWithinBand_largeFont() {
-        assertRecordingRowFits(locked = true, fontScale = 1.3f)
-    }
-
-    private fun assertRecordingRowFits(locked: Boolean, fontScale: Float) {
+    private fun assertRecordingRowFits(fontScale: Float) {
         compose.activityRule.scenario.onActivity { activity ->
             WindowCompat.setDecorFitsSystemWindows(activity.window, false)
         }
@@ -128,7 +104,6 @@ class PromptComposerRecordingRowFitProofTest {
         val state = PromptComposerViewModel.UiState(
             draft = "",
             recording = PromptComposerViewModel.RecordingState.Recording,
-            recordingLocked = locked,
             amplitude = 0.5f,
             hasDetectedSpeech = true,
             recordingElapsedMs = 14_000L,
@@ -170,13 +145,12 @@ class PromptComposerRecordingRowFitProofTest {
                                 onDraftChange = {},
                                 onMicTap = {},
                                 onSend = {},
-                                // Non-null so the editing tools + snippets render
-                                // (they must stay MOUNTED during recording, and
-                                // must fit — the maintainer directive).
+                                // Non-null so the editing tools + snippets are
+                                // available in the Idle/Transcribing states; during
+                                // Recording they are intentionally not shown.
                                 onAttachFiles = {},
                                 onSnippets = {},
                                 onCancelRecording = {},
-                                onLockRecording = {},
                             )
                         }
                     }
@@ -205,24 +179,17 @@ class PromptComposerRecordingRowFitProofTest {
         val slopPx = CONTAINMENT_SLOP_DP * density
         val band = bounds(BAND_TAG)
 
-        // The pills that MUST be present + fully within the band in this state,
-        // each with a not-a-sliver minimum width. A labelled pill's natural width
-        // is 70–102dp; base crushes the trailing `Send` to ~33dp, so a 56dp floor
-        // cleanly separates crushed (red) from intact (green). The editing-tools
-        // group is a fixed 3×40dp icon pill, so it uses the icon-width floor.
-        // Issue #1245: the hands-free Lock is no longer a bottom-cluster pill — it
-        // moved UP to the recording/waveform row as an inline toggle, so it is not
-        // part of this bottom-row fit check (its placement is proven by
-        // PromptComposerMicSwipeLockJourneyTest). The bottom row must still FIT
-        // Discard · Insert · Send with the tools group mounted (the #1152 guard).
-        val pills = buildList {
-            add(Triple("Discard", COMPOSER_CANCEL_RECORDING_TAG, LABELLED_PILL_MIN_DP))
-            add(Triple("Insert", COMPOSER_TO_FIELD_TAG, LABELLED_PILL_MIN_DP))
-            add(Triple("Send", COMPOSER_STOP_SEND_TAG, LABELLED_PILL_MIN_DP))
-            // The editing tools group stays mounted during recording (the "fit
-            // everything, don't hide" directive) — prove it is present + fits too.
-            add(Triple("Tools(attach)", COMPOSER_ATTACH_TAG, ICON_MIN_DP))
-        }
+        // The three stop-action pills that MUST be present + fully within the band
+        // in the recording state, each with a not-a-sliver minimum width. A labelled
+        // pill's natural width is 70–102dp; a crushed sliver is ~33dp, so a 56dp
+        // floor cleanly separates crushed (red) from intact (green). Issue #1245:
+        // the hands-free Lock is gone and the editing-tools group is not shown
+        // mid-dictation, so the bottom row is exactly [Discard · Insert · Send].
+        val pills = listOf(
+            Triple("Discard", COMPOSER_CANCEL_RECORDING_TAG, LABELLED_PILL_MIN_DP),
+            Triple("Insert", COMPOSER_TO_FIELD_TAG, LABELLED_PILL_MIN_DP),
+            Triple("Send", COMPOSER_STOP_SEND_TAG, LABELLED_PILL_MIN_DP),
+        )
 
         for ((label, tag, minDp) in pills) {
             val pill = bounds(tag)
@@ -232,51 +199,36 @@ class PromptComposerRecordingRowFitProofTest {
             val heightDp = (pill.bottom - pill.top) / density
             // Log geometry FIRST so a failure is diagnosable straight from the log.
             println(
-                "ISSUE1152_FIT locked=$locked font=$fontScale $label " +
+                "ISSUE1152_FIT font=$fontScale $label " +
                     "pill=[${pill.left}..${pill.right}] band=[${band.left}..${band.right}] " +
                     "widthDp=$widthDp heightDp=$heightDp minDp=$minDp " +
                     "overflowRightPx=$overflowRightPx overflowLeftPx=$overflowLeftPx",
             )
-            // LOAD-BEARING: the pill must not be crushed to a sliver. On base the
-            // un-weighted overflow squeezes `Send` to ~33dp (the "S/e" sliver);
-            // the two-row fit keeps every pill at its full width.
+            // LOAD-BEARING: the pill must not be crushed to a sliver. A crushed pill
+            // still reports assertIsDisplayed() (the #657/F1 trap), so this width
+            // floor is the real check.
             assertTrue(
-                "Recording-row pill '$label' is CRUSHED to a sliver (audit D1 — " +
-                    "the overflow squeezes it below a legible width). widthDp=$widthDp " +
-                    "minDp=$minDp locked=$locked font=$fontScale. A crushed pill still " +
-                    "reports assertIsDisplayed() (the #657/F1 trap), so this width " +
-                    "floor is the real check.",
+                "Recording-row pill '$label' is CRUSHED to a sliver (audit D1). " +
+                    "widthDp=$widthDp minDp=$minDp font=$fontScale. A crushed pill " +
+                    "still reports assertIsDisplayed(), so this width floor is the " +
+                    "real check.",
                 widthDp >= minDp,
             )
             assertTrue(
                 "Recording-row pill '$label' spills off the RIGHT of the band " +
                     "(clipped/off-edge — audit D1). pill.right=${pill.right} " +
                     "band.right=${band.right} overflowRightPx=$overflowRightPx " +
-                    "widthDp=$widthDp locked=$locked font=$fontScale",
+                    "widthDp=$widthDp font=$fontScale",
                 pill.right <= band.right + slopPx,
             )
             assertTrue(
                 "Recording-row pill '$label' spills off the LEFT of the band. " +
                     "pill.left=${pill.left} band.left=${band.left} " +
-                    "overflowLeftPx=$overflowLeftPx locked=$locked font=$fontScale",
+                    "overflowLeftPx=$overflowLeftPx font=$fontScale",
                 pill.left >= band.left - slopPx,
             )
             // Additive brief-named guard: never off the WHOLE screen either.
             compose.assertNodeFullyWithinRoot(tag)
-        }
-
-        // Issue #1245: the inline hands-free lock toggle now rides the recording/
-        // waveform row (when not locked). Guard that it, too, stays fully within
-        // the band (never spills off the right edge into the waveform's neighbour
-        // area) at both font scales.
-        if (!locked) {
-            val lock = bounds(COMPOSER_LOCK_RECORDING_TAG)
-            assertTrue(
-                "Inline lock toggle spills off the band. lock=[${lock.left}..${lock.right}] " +
-                    "band=[${band.left}..${band.right}] font=$fontScale",
-                lock.right <= band.right + slopPx && lock.left >= band.left - slopPx,
-            )
-            compose.assertNodeFullyWithinRoot(COMPOSER_LOCK_RECORDING_TAG)
         }
     }
 
@@ -316,10 +268,9 @@ class PromptComposerRecordingRowFitProofTest {
     private companion object {
         const val BAND_TAG = "issue1152-recording-band"
 
-        // A real phone content width (Pixel-class, matches #780's 392dp). The old
-        // single-row content (~421dp) overflows it, so BASE (pre-fix) is RED here;
-        // the two-row fix fits with comfortable margin (secondary row ~330dp,
-        // primary row ~215dp) at both 1.0 and 1.3 font scale.
+        // A real phone content width (Pixel-class, matches #780's 392dp). The three
+        // stop pills [Discard · Insert · Send] fit with comfortable margin at both
+        // 1.0 and 1.3 font scale.
         const val BAND_WIDTH_DP = 392f
         const val BAND_HEIGHT_DP = 760f
 
@@ -329,11 +280,9 @@ class PromptComposerRecordingRowFitProofTest {
 
         const val CONTAINMENT_SLOP_DP = 1f
 
-        // A labelled recording pill's natural width is 70–102dp (Discard/Lock/
-        // Insert/Send) across font scale 1.0–1.3; base crushes the trailing `Send`
-        // to ~33dp. 56dp is the not-a-sliver floor between them.
+        // A labelled recording pill's natural width is 70–102dp (Discard/Insert/
+        // Send) across font scale 1.0–1.3; a crushed pill is ~33dp. 56dp is the
+        // not-a-sliver floor between them.
         const val LABELLED_PILL_MIN_DP = 56f
-        // The editing-tools group is a fixed 3×40dp icon pill; 36dp is its floor.
-        const val ICON_MIN_DP = 36f
     }
 }
