+++ app/src/main/java/com/pocketshell/app/composer/PromptComposerViewModel.kt
@@ -786,29 +786,6 @@ public class PromptComposerViewModel @Inject constructor(
     }
 
     /**
-     * Issue #585: mark the active dictation as swipe-up locked. The lock does
-     * not change how the recorder is stopped: the explicit Recording controls
-     * still decide whether to discard, insert, or send. It gives the UI an
-     * observable state for the Telegram-style mic gesture without affecting a
-     * normal tap-to-record flow.
-     */
-    public fun lockRecording() {
-        if (_uiState.value.recording != RecordingState.Recording) return
-        var lockedNow = false
-        _uiState.update { current ->
-            if (current.recordingLocked) {
-                current
-            } else {
-                lockedNow = true
-                current.copy(recordingLocked = true)
-            }
-        }
-        if (lockedNow) {
-            DiagnosticEvents.record("action", "composer_recording_lock")
-        }
-    }
-
-    /**
      * Issue #211 / #745: emit a [SendRequest] for the current draft and
      * mark the send IN-FLIGHT — the draft and any staged attachments are
      * deliberately NOT cleared here. The composer keeps the typed text
@@ -1222,7 +1199,6 @@ public class PromptComposerViewModel @Inject constructor(
                 current.copy(
                     recording = RecordingState.Idle,
                     amplitude = 0f,
-                    recordingLocked = false,
                     sendInFlight = false,
                     attachmentUpload = AttachmentUploadState.Idle,
                 )
@@ -1249,7 +1225,6 @@ public class PromptComposerViewModel @Inject constructor(
                 error = message,
                 recording = RecordingState.Idle,
                 amplitude = 0f,
-                recordingLocked = false,
                 sendInFlight = false,
                 // Issue #872: re-show the actual attachment tiles so Retry
                 // re-sends the original attachment, not just the text.
@@ -1406,7 +1381,6 @@ public class PromptComposerViewModel @Inject constructor(
             it.copy(
                 recording = RecordingState.Idle,
                 amplitude = 0f,
-                recordingLocked = false,
                 error = TRANSCRIBE_TIMEOUT_MESSAGE,
             )
         }
@@ -2036,7 +2010,6 @@ public class PromptComposerViewModel @Inject constructor(
             _uiState.update {
                 it.copy(
                     recording = RecordingState.Idle,
-                    recordingLocked = false,
                     error = e.message ?: "Microphone unavailable",
                 )
             }
@@ -2077,7 +2050,6 @@ public class PromptComposerViewModel @Inject constructor(
             it.copy(
                 recording = RecordingState.Recording,
                 amplitude = 0f,
-                recordingLocked = false,
                 // Issue #195: every fresh recording starts in the
                 // "listening, no speech yet" sub-state. The sampling loop
                 // below flips this to true on the first amplitude crossing.
@@ -2173,7 +2145,6 @@ public class PromptComposerViewModel @Inject constructor(
             it.copy(
                 recording = RecordingState.Recording,
                 amplitude = 0.12f,
-                recordingLocked = false,
                 hasDetectedSpeech = false,
                 silenceThresholdSeconds = 0f,
                 recordingElapsedMs = 0L,
@@ -2337,7 +2308,6 @@ public class PromptComposerViewModel @Inject constructor(
                 it.copy(
                     recording = RecordingState.Idle,
                     amplitude = 0f,
-                    recordingLocked = false,
                     error = e.message ?: "Microphone error",
                 )
             }
@@ -2362,7 +2332,7 @@ public class PromptComposerViewModel @Inject constructor(
                 "audioBytes" to 0,
             )
             _uiState.update {
-                it.copy(recording = RecordingState.Idle, amplitude = 0f, recordingLocked = false)
+                it.copy(recording = RecordingState.Idle, amplitude = 0f)
             }
             return
         }
@@ -2388,7 +2358,7 @@ public class PromptComposerViewModel @Inject constructor(
             )
             if (recoverableNoSpeech) {
                 _uiState.update {
-                    it.copy(recording = RecordingState.Transcribing, amplitude = 0f, recordingLocked = false)
+                    it.copy(recording = RecordingState.Transcribing, amplitude = 0f)
                 }
                 // Issue #939 (#935 S5-2): even this recoverable-no-speech path
                 // flips the FSM to Transcribing then does an UNGUARDED Room write
@@ -2406,7 +2376,6 @@ public class PromptComposerViewModel @Inject constructor(
                         it.copy(
                             recording = RecordingState.Idle,
                             amplitude = 0f,
-                            recordingLocked = false,
                             error = if (pendingId != null) {
                                 SUSPICIOUS_NO_SPEECH_RETRY_MESSAGE
                             } else {
@@ -2424,7 +2393,6 @@ public class PromptComposerViewModel @Inject constructor(
                 it.copy(
                     recording = RecordingState.Idle,
                     amplitude = 0f,
-                    recordingLocked = false,
                     error = NO_SPEECH_DETECTED_MESSAGE,
                 )
             }
@@ -2440,7 +2408,7 @@ public class PromptComposerViewModel @Inject constructor(
             "audioBytes" to audio.size,
         )
         _uiState.update {
-            it.copy(recording = RecordingState.Transcribing, amplitude = 0f, recordingLocked = false)
+            it.copy(recording = RecordingState.Transcribing, amplitude = 0f)
         }
 
         launchTranscribe {
@@ -2464,7 +2432,6 @@ public class PromptComposerViewModel @Inject constructor(
                 _uiState.update {
                     it.copy(
                         recording = RecordingState.Idle,
-                        recordingLocked = false,
                         error = "No OpenAI API key saved. Open settings to add one.",
                     )
                 }
@@ -2527,7 +2494,6 @@ public class PromptComposerViewModel @Inject constructor(
                 _uiState.update {
                     it.copy(
                         recording = RecordingState.Idle,
-                        recordingLocked = false,
                         error = PendingTranscriptionItem.NETWORK_WAITING_MESSAGE,
                     )
                 }
@@ -2567,7 +2533,6 @@ public class PromptComposerViewModel @Inject constructor(
                             it.copy(
                                 recording = RecordingState.Idle,
                                 amplitude = 0f,
-                                recordingLocked = false,
                                 error = if (pendingId != null) {
                                     EMPTY_TRANSCRIPTION_RETRY_MESSAGE
                                 } else {
@@ -2608,7 +2573,6 @@ public class PromptComposerViewModel @Inject constructor(
                             it.copy(
                                 recording = RecordingState.Idle,
                                 amplitude = 0f,
-                                recordingLocked = false,
                                 error = NO_SPEECH_DETECTED_MESSAGE,
                             )
                         }
@@ -2683,7 +2647,6 @@ public class PromptComposerViewModel @Inject constructor(
                         it.copy(
                             recording = RecordingState.Idle,
                             amplitude = 0f,
-                            recordingLocked = false,
                             error = msg,
                         )
                     }
@@ -2749,7 +2712,6 @@ public class PromptComposerViewModel @Inject constructor(
                     it.copy(
                         recording = RecordingState.Idle,
                         amplitude = 0f,
-                        recordingLocked = false,
                         error = userFacingWhisperError(t),
                     )
                 }
@@ -2763,7 +2725,6 @@ public class PromptComposerViewModel @Inject constructor(
                         it.copy(
                             recording = RecordingState.Idle,
                             amplitude = 0f,
-                            recordingLocked = false,
                         )
                     }
                 }
@@ -2807,7 +2768,6 @@ public class PromptComposerViewModel @Inject constructor(
             it.copy(
                 recording = RecordingState.Transcribing,
                 amplitude = 0f,
-                recordingLocked = false,
                 error = null,
             )
         }
@@ -2864,7 +2824,6 @@ public class PromptComposerViewModel @Inject constructor(
                 recording = RecordingState.Idle,
                 amplitude = 0f,
                 hasDetectedSpeech = false,
-                recordingLocked = false,
                 liveTranscript = null,
                 error = null,
             )
@@ -2895,7 +2854,6 @@ public class PromptComposerViewModel @Inject constructor(
                 recording = RecordingState.Idle,
                 amplitude = 0f,
                 hasDetectedSpeech = false,
-                recordingLocked = false,
                 liveTranscript = null,
                 error = androidSpeechFailureMessage(message),
             )
@@ -2919,7 +2877,6 @@ public class PromptComposerViewModel @Inject constructor(
                 recording = RecordingState.Idle,
                 amplitude = 0f,
                 hasDetectedSpeech = false,
-                recordingLocked = false,
                 liveTranscript = null,
                 draft = restoredDraft,
                 error = null,
@@ -2977,7 +2934,6 @@ public class PromptComposerViewModel @Inject constructor(
             it.copy(
                 recording = RecordingState.Idle,
                 draft = newDraft,
-                recordingLocked = false,
                 error = null,
             )
         }
@@ -3282,7 +3238,6 @@ public class PromptComposerViewModel @Inject constructor(
             it.copy(
                 recording = RecordingState.Idle,
                 amplitude = 0f,
-                recordingLocked = false,
                 // Existing typed draft must survive. The user's intent
                 // here is "throw away the new dictation, keep what I
                 // already typed" — wiping the draft would be a hostile
@@ -3327,7 +3282,6 @@ public class PromptComposerViewModel @Inject constructor(
             it.copy(
                 recording = RecordingState.Idle,
                 amplitude = 0f,
-                recordingLocked = false,
                 error = null,
             )
         }
@@ -3342,7 +3296,6 @@ public class PromptComposerViewModel @Inject constructor(
         _uiState.update {
             it.copy(
                 recording = RecordingState.Idle,
-                recordingLocked = false,
                 error = "Microphone permission denied. Grant it in system settings to use voice input.",
             )
         }
@@ -3635,17 +3588,12 @@ public class PromptComposerViewModel @Inject constructor(
      *   after Xs silence" hint so the user has a mental model. Zero while
      *   the FSM is in [RecordingState.Idle] / [RecordingState.Transcribing]
      *   so the hint is hidden outside an active recording.
-     * @param recordingLocked issue #585: true when the user swiped the mic
-     *   upward past the lock threshold during this active recording. It is a
-     *   UI affordance only; the explicit stop/discard controls still decide
-     *   how the captured audio is handled.
      */
     public data class UiState(
         val draft: String = "",
         val recording: RecordingState = RecordingState.Idle,
         val amplitude: Float = 0f,
         val hasDetectedSpeech: Boolean = false,
-        val recordingLocked: Boolean = false,
         val error: String? = null,
         val savedAudioPath: String? = null,
         val silenceThresholdSeconds: Float = 0f,
