+++ app/src/test/java/com/pocketshell/app/composer/PromptComposerViewModelTest.kt
@@ -1640,53 +1640,23 @@ class PromptComposerViewModelTest {
         assertNull(vm.uiState.value.error)
     }
 
-    // -- Issue #585: mic swipe-up lock ---------------------------------------
+    // -- Issue #1245: recording works with NO lock concept -------------------
 
     @Test
-    fun lockRecordingOnlyAppliesDuringActiveRecording() = runTest {
+    fun recordingRunsWithoutAnyLockConceptUntilStopped() = runTest {
         val vm = newVm(samplerDispatcher = StandardTestDispatcher(testScheduler))
 
-        vm.lockRecording()
-        assertFalse(vm.uiState.value.recordingLocked)
-
+        // A plain mic tap starts recording — there is no press-and-hold or
+        // swipe-to-lock step; recording simply runs.
         vm.onMicTap()
         runCurrent()
-        vm.lockRecording()
         assertEquals(RecordingState.Recording, vm.uiState.value.recording)
-        assertTrue(vm.uiState.value.recordingLocked)
 
+        // The recording keeps running (Discard / Insert / Send are the only ways
+        // to end it), and cancel returns to Idle cleanly.
         vm.cancelRecording()
         advanceUntilIdle()
         assertEquals(RecordingState.Idle, vm.uiState.value.recording)
-        assertFalse(vm.uiState.value.recordingLocked)
-    }
-
-    @Test
-    fun stoppingLockedRecordingClearsLockBeforeTranscribing() = runTest {
-        val release = CompletableDeferred<Unit>()
-        val whisper = object : WhisperClient {
-            override suspend fun transcribe(audio: ByteArray, language: String?): Result<String> {
-                release.await()
-                return Result.success("locked words")
-            }
-        }
-        val vm = newVm(
-            whisper = whisper,
-            samplerDispatcher = StandardTestDispatcher(testScheduler),
-        )
-
-        vm.onMicTap()
-        runCurrent()
-        vm.lockRecording()
-        assertTrue(vm.uiState.value.recordingLocked)
-
-        vm.onMicTap()
-        runCurrent()
-        assertEquals(RecordingState.Transcribing, vm.uiState.value.recording)
-        assertFalse(vm.uiState.value.recordingLocked)
-
-        release.complete(Unit)
-        advanceUntilIdle()
     }
 
     // -- Issue #174: cancel-recording affordance -----------------------------
